AI Agents/requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langchain-groq>=0.0.1
crewai>=0.22.0
pandas>=2.0.0
PyPDF2>=3.0.0
pillow>=10.0.0
opencv-python>=4.8.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
sqlalchemy>=2.0.0
pyyaml>=6.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0
spacy>=3.7.0
pyspellchecker>=0.8.0

